requirements.txt
atproto==0.0.41
requests==2.31.0
feedparser==6.0.10
newspaper3k==0.2.8
beautifulsoup4==4.12.2
google-genai==0.3.0
python-dotenv==1.0.0